requirements.txt
dspy_ai==2.4.9
fastapi==0.112.2
gunicorn==23.0.0
knowledge_storm==0.2.4
multiprocess==0.70.15
openai==1.40.6
passlib==1.7.4
PyJWT==2.9.0
pydantic==2.8.2
pydantic_settings==2.4.0
Requests==2.32.3
redis==5.0.8
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.38.2
uvicorn==0.30.6